% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-raster.R
\name{distance_raster}
\alias{distance_raster}
\title{Distance raster}
\usage{
distance_raster(y, cellsize, extent = NULL, expand = NULL, check = TRUE)
}
\arguments{
\item{cellsize}{in unit of projection}

\item{extent}{optional alternative extent bbox}

\item{expand}{0-1 scaling eg. 5\% expansion = 0.05}
}
\value{

}
\description{
Warnings + recommend pt based
}
